fix: PR Comment 반영
CombinationPreviewDTO hashTag 정보 추가

Related: #26
