+++ src/main/java/com/example/dgbackend/domain/combinationcomment/domain/CombinationComment.java
@@ -1,7 +1,7 @@
-package com.example.dgbackend.domain.combination;
+package com.example.dgbackend.domain.combinationcomment.domain;
 
-
-import com.example.dgbackend.domain.member.Member;
+import com.example.dgbackend.domain.combination.domain.Combination;
+import com.example.dgbackend.domain.member.domain.Member;
 import com.example.dgbackend.global.common.BaseTimeEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -11,38 +11,34 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import org.hibernate.annotations.ColumnDefault;
 
+@Builder
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class Combination extends BaseTimeEntity {
+public class CombinationComment extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @NotNull
-    private String title;
-
-    @NotNull
     private String content;
 
     @ColumnDefault("0")
-    private Long likeCount;
-
-    @ColumnDefault("0")
-    private Long commentCount;
+    private Long parentId; //댓글 : 0, 대 댓글 : 자신의 부모 댓글 id
 
     private boolean state = true; //true : 존재, false : 삭제
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "combination_id")
+    private Combination combination;
 
 }
