+++ src/main/java/com/example/dgbackend/domain/combinationimage/CombinationImage.java
@@ -1,6 +1,6 @@
 package com.example.dgbackend.domain.combinationimage;
 
-import com.example.dgbackend.domain.combination.Combination;
+import com.example.dgbackend.domain.combination.domain.Combination;
 import com.example.dgbackend.global.common.BaseTimeEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
