+++ src/main/java/com/example/dgbackend/domain/recipecomment/RecipeComment.java
@@ -1,7 +1,7 @@
 package com.example.dgbackend.domain.recipecomment;
 
 import com.example.dgbackend.domain.recipe.Recipe;
-import com.example.dgbackend.domain.member.Member;
+import com.example.dgbackend.domain.member.domain.Member;
 import com.example.dgbackend.global.common.BaseTimeEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
