+++ src/main/java/com/example/dgbackend/domain/recommend/Recommend.java
@@ -1,6 +1,6 @@
 package com.example.dgbackend.domain.recommend;
 
-import com.example.dgbackend.domain.member.Member;
+import com.example.dgbackend.domain.member.domain.Member;
 import com.example.dgbackend.global.common.BaseTimeEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
